// packages/test/src/protos-es-gen/namespaced-messages_pb.ts
// @generated by protoc-gen-es v2.12.0 with parameter "target=ts,import_extension=none"
// @generated from file namespaced-messages.proto (package foo.bar, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file namespaced-messages.proto.
 */
export const file_namespaced_messages: GenFile = /*@__PURE__*/
  fileDesc("ChluYW1lc3BhY2VkLW1lc3NhZ2VzLnByb3RvEgdmb28uYmFyIi8KElByb3RvQWN0aXZpdHlJbnB1dBIMCgRuYW1lGAEgASgJEgsKA2FnZRgCIAEoBWIGcHJvdG8z");

/**
 * @generated from message foo.bar.ProtoActivityInput
 */
export type ProtoActivityInput = Message<"foo.bar.ProtoActivityInput"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: int32 age = 2;
   */
  age: number;
};

/**
 * Describes the message foo.bar.ProtoActivityInput.
 * Use `create(ProtoActivityInputSchema)` to create a new message.
 */
export const ProtoActivityInputSchema: GenMessage<ProtoActivityInput> = /*@__PURE__*/
  messageDesc(file_namespaced_messages, 0);